Combine crypto unittests into crypto_test.cc
